fix: corrected all comments
