+++ Infrastructure/BankSystem.Data/Extensions/EnumerableExtensions.cs
@@ -1,6 +1,5 @@
-namespace BankSystem.Domain;
+namespace BankSystem.Data.Extensions;
 
-// не уверен, что это следует размещать здесь, надеюсь подскажите комментарием
 public static class EnumerableExtensions
 {
     public static IEnumerable<T> WhereIf<T>(this IEnumerable<T> source, bool condition, Func<T, bool> predicate)
